+++ src/main/java/lotto/Application.java
@@ -3,13 +3,12 @@ package lotto;
 import lotto.controller.GameController;
 import lotto.controller.InputController;
 
-import lotto.domain.service.Money;
-import lotto.domain.service.User;
+import lotto.domain.model.Money;
 import lotto.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        GameController gameController = new GameController(new InputController(new Money()),new OutputView(),new User(),new Money());
+        GameController gameController = new GameController(new InputController(),new OutputView());
         gameController.startGame();
         gameController.showResult();
     }
